add QA stories
