+++ stories/qa/forms/checkbox-legacy.stories.html
@@ -1,5 +1,5 @@
 <section class="contentSection">
-	<h1>Checkboxes</h1>
+	<h1>Checkbox Legacy</h1>
 
 	<!-- Basics -->
 	<h2>Basic</h2>
