+++ stories/qa/forms/checkbox.stories.html
@@ -1,0 +1,164 @@
+<section class="contentSection">
+	<h1>Checkbox field</h1>
+
+	<h2>Basic</h2>
+	<div class="checkboxField">
+		<input type="checkbox" class="checkboxField-input" id="CB1" aria-labelledby="CB1label" aria-describedby="CB1message" aria-required="true" />
+		<label class="checkboxField-label" for="CB1">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB1label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB1message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+	<div class="checkboxField">
+		<input type="checkbox" class="checkboxField-input" id="CB2" aria-labelledby="CB2label" aria-describedby="CB2message" aria-required="true" checked />
+		<label class="checkboxField-label" for="CB2">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB2label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB2message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+	<div class="checkboxField">
+		<input type="checkbox" class="checkboxField-input" id="CB3" aria-labelledby="CB3label" aria-describedby="CB3message" aria-required="true" disabled />
+		<label class="checkboxField-label" for="CB3">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB3label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB3message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+	<div class="checkboxField">
+		<input type="checkbox" class="checkboxField-input" id="CB4" aria-labelledby="CB4label" aria-describedby="CB4message" aria-required="true" disabled checked />
+		<label class="checkboxField-label" for="CB4">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB4label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB4message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+
+	<h2>Small</h2>
+	<div class="checkboxField mod-S">
+		<input type="checkbox" class="checkboxField-input" id="CB5" aria-labelledby="CB5label" aria-describedby="CB5message" aria-required="true" />
+		<label class="checkboxField-label" for="CB5">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB5label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB5message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+	<div class="checkboxField mod-S">
+		<input type="checkbox" class="checkboxField-input" id="CB6" aria-labelledby="CB6label" aria-describedby="CB6message" aria-required="true" checked />
+		<label class="checkboxField-label" for="CB6">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB6label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB6message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+
+	<h2>Mixed</h2>
+	<div class="checkboxField">
+		<input type="checkbox" class="checkboxField-input" id="CB7" aria-labelledby="CB7label" aria-describedby="CB7message" aria-required="true" aria-checked="mixed" checked />
+		<label class="checkboxField-label" for="CB7">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB7label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB7message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+	<div class="checkboxField mod-S">
+		<input type="checkbox" class="checkboxField-input" id="CB8" aria-labelledby="CB8label" aria-describedby="CB8message" aria-required="true" aria-checked="mixed" checked />
+		<label class="checkboxField-label" for="CB8">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB8label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB8message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+
+	<h2>Invalid</h2>
+	<div class="checkboxField">
+		<input type="checkbox" class="checkboxField-input" id="CB9" aria-labelledby="CB9label" aria-describedby="CB9message" aria-required="true" aria-invalid="true" checked />
+		<label class="checkboxField-label" for="CB9">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB9label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB9message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+	<div class="checkboxField mod-S">
+		<input type="checkbox" class="checkboxField-input" id="CB10" aria-labelledby="CB10label" aria-describedby="CB10message" aria-required="true" aria-invalid="true" checked />
+		<label class="checkboxField-label" for="CB10">
+			<span class="checkboxField-label-input">
+				<span class="checkboxField-label-input-icon" aria-hidden="true"></span>
+			</span>
+			<span class="checkboxField-label-text" id="CB10label">
+				Label<sup aria-hidden="true" class="checkboxField-label-text-required">*</sup>
+			</span>
+		</label>
+		<div class="checkboxField-message" id="CB10message">
+			<span class="checkboxField-message-icon" aria-hidden="true"></span>
+			<span class="checkboxField-message-text">Helper text</span>
+		</div>
+	</div>
+</section>
+
+<!-- To tell the ui-diff tool that the page has finished rendering -->
+<span id="ready"></span>
