--- stories/qa/forms/checkboxes.stories.ts
@@ -1,22 +1,0 @@
-import { Component } from '@angular/core';
-import { Story, Meta, moduleMetadata } from '@storybook/angular';
-
-@Component({
-	standalone: true,
-	selector: 'forms-checkboxes-stories',
-	templateUrl: './checkboxes.stories.html'
-}) class CheckboxesStory {}
-
-export default {
-  title: 'QA/Forms/Checkboxes',
-  component: CheckboxesStory,
-	decorators: [
-		moduleMetadata({
-			entryComponents: [CheckboxesStory]
-		})
-	]
-} as Meta;
-
-const template: Story<CheckboxesStory> = () => ({});
-
-export const basic = template.bind({});
